+++ src/bin_node/node_config_file.ml
@@ -353,6 +353,7 @@ and p2p = {
   expected_pow : float;
   bootstrap_peers : string list option;
   listen_addr : string option;
+  advertised_net_port : int option;
   discovery_addr : string option;
   private_mode : bool;
   limits : P2p.limits;
@@ -415,6 +416,7 @@ let default_p2p =
     expected_pow = 26.;
     bootstrap_peers = None;
     listen_addr = Some ("[::]:" ^ string_of_int default_p2p_port);
+    advertised_net_port = None;
     discovery_addr = None;
     private_mode = false;
     limits = default_p2p_limits;
@@ -675,6 +677,7 @@ let p2p =
            expected_pow;
            bootstrap_peers;
            listen_addr;
+           advertised_net_port;
            discovery_addr;
            private_mode;
            limits;
@@ -685,6 +688,7 @@ let p2p =
       ( expected_pow,
         bootstrap_peers,
         listen_addr,
+        advertised_net_port,
         discovery_addr,
         private_mode,
         limits,
@@ -694,6 +698,7 @@ let p2p =
     (fun ( expected_pow,
            bootstrap_peers,
            listen_addr,
+           advertised_net_port,
            discovery_addr,
            private_mode,
            limits,
@@ -704,6 +709,7 @@ let p2p =
         expected_pow;
         bootstrap_peers;
         listen_addr;
+        advertised_net_port;
         discovery_addr;
         private_mode;
         limits;
@@ -711,7 +717,7 @@ let p2p =
         enable_testchain;
         reconnection_config;
       })
-    (obj9
+    (obj10
        (dft
           "expected-proof-of-work"
           ~description:
@@ -732,6 +738,12 @@ let p2p =
             "Host to listen to. If the port is not specified, the default port \
              9732 will be assumed."
           string)
+       (opt
+          "advertised-net-port"
+          ~description:
+            "Alternative port advertised to other peers to connect to. If the \
+             port is not specified, the port from listen-addr will be assumed."
+          uint16)
        (dft
           "discovery-addr"
           ~description:
@@ -1200,8 +1212,8 @@ let to_string cfg =
 let update ?(disable_config_validation = false) ?data_dir ?min_connections
     ?expected_connections ?max_connections ?max_download_speed ?max_upload_speed
     ?binary_chunks_size ?peer_table_size ?expected_pow ?bootstrap_peers
-    ?listen_addr ?discovery_addr ?(rpc_listen_addrs = []) ?(allow_all_rpc = [])
-    ?(private_mode = false) ?(disable_mempool = false)
+    ?listen_addr ?advertised_net_port ?discovery_addr ?(rpc_listen_addrs = [])
+    ?(allow_all_rpc = []) ?(private_mode = false) ?(disable_mempool = false)
     ?(enable_testchain = false) ?(cors_origins = []) ?(cors_headers = [])
     ?rpc_tls ?log_output ?synchronisation_threshold ?history_mode ?network
     ?latency cfg =
@@ -1252,6 +1264,8 @@ let update ?(disable_config_validation = false) ?data_dir ?min_connections
       bootstrap_peers =
         Option.value ~default:cfg.p2p.bootstrap_peers bootstrap_peers;
       listen_addr = Option.either listen_addr cfg.p2p.listen_addr;
+      advertised_net_port =
+        Option.either advertised_net_port cfg.p2p.advertised_net_port;
       discovery_addr = Option.either discovery_addr cfg.p2p.discovery_addr;
       private_mode = cfg.p2p.private_mode || private_mode;
       limits;